Fix a bug in dcpu_skip().  The 'b' instruction operand (or 'a' operand for non-basic opcodes) could overflow skiptable, when used as an index. Masking the operand to force it within the table size has its own problem, as it causes aliasing from some constant operands to operands which "advance pc".
Example:

c00d	009c:	IFN A, 16
fc03	009d:	SUB A, 31

The constant 31 in the second instruction is 'b' operand value 0x3F.  If
masked with 31 (0x1F) to stay within the table, it becomes 0x1F, which has
value 1 within skiptable - that's incorrect, the constant does not require
another instruction word.

The solution in this commit tries to perturb the existing code as little
as possible; I just bump skiptable up to 64-locations, so there's no
need to mask the operand before using it as an index.  The other change
is to update the mask on the basic-opcode 'a' operand so that literals
are handled correcty for that operand as well.

I've included a test case, "testskip3.s", which fails before the fix and
passes after the fix.  All 32 literal operand values for the 'b' operand
are tested.